# clawgotcha/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "clawgotcha",
    "version": "1.0.0",
    "description": "HTTP API for MiroClaw/ZeroClaw-style swarm agent metadata: default provider/model, Hands ([[agents]]), and cron jobs ([[cron_jobs]]) backed by PostgreSQL. **Errors** return JSON `{\"error\":{\"code\":\"...\",\"detail\":\"...\"}}` (see `Error`). `4xx` and `5xx` use the same error envelope."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3477",
      "description": "Default local listen (see `port` in YAML, default 3477; or `HTTP_ADDR`)"
    }
  ],
  "tags": [
    { "name": "Health" },
    { "name": "Meta" },
    { "name": "Config" },
    { "name": "Agents" },
    { "name": "Cron jobs" },
    { "name": "Instances" },
    { "name": "Events" }
  ],
  "paths": {
    "/healthz": {
      "get": {
        "tags": ["Health"],
        "summary": "Liveness",
        "responses": {
          "200": {
            "description": "Service is up",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Healthz" }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "tags": ["Meta"],
        "summary": "This OpenAPI document",
        "responses": {
          "200": {
            "description": "OpenAPI 3 JSON"
          }
        }
      }
    },
    "/api/v1/config": {
      "get": {
        "tags": ["Config"],
        "summary": "Get swarm defaults (single row, id=1)",
        "responses": {
          "200": {
            "description": "Default provider and model",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmConfig" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "tags": ["Config"],
        "summary": "Replace swarm defaults",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PutConfigRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated config",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmConfig" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/agents": {
      "get": {
        "tags": ["Agents"],
        "summary": "List agents (Hands), ordered by name",
        "responses": {
          "200": {
            "description": "List wrapper",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AgentListResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Agents"],
        "summary": "Create an agent (Hand)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrReplaceAgentRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created agent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmAgent" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/agents/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "get": {
        "tags": ["Agents"],
        "summary": "Get agent by id",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmAgent" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "tags": ["Agents"],
        "summary": "Replace agent by id (full body; same fields as create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrReplaceAgentRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmAgent" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "tags": ["Agents"],
        "summary": "Partially update an agent. If `clear_tools` is true, `tools` is set to []. Omitted fields are unchanged; empty `updates` returns current row.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatchAgentRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmAgent" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["Agents"],
        "summary": "Delete agent by id",
        "responses": {
          "204": { "description": "Deleted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/agents/by-name/{name}": {
      "get": {
        "tags": ["Agents"],
        "summary": "Get agent by unique name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmAgent" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/cron-jobs/schedule-timeline": {
      "get": {
        "tags": ["Cron jobs"],
        "summary": "List cron jobs with projected run times (standard 5-field cron) anchored on `UpdatedAt` and filtered by `Active` & horizon",
        "parameters": [
          {
            "name": "horizon_hours",
            "in": "query",
            "description": "Window length from server `as_of` (default 168, max 8760)",
            "schema": { "type": "integer", "default": 168, "minimum": 1, "maximum": 8760 }
          },
          {
            "name": "max_runs",
            "in": "query",
            "description": "Cap of projected instants per job (default 64, max 200)",
            "schema": { "type": "integer", "default": 64, "minimum": 1, "maximum": 200 }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CronScheduleTimelineResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/cron-jobs": {
      "get": {
        "tags": ["Cron jobs"],
        "summary": "List cron jobs, ordered by name",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CronJobListResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Cron jobs"],
        "summary": "Create a cron job. `agent_name` must match an existing agent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrReplaceCronJobRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmCronJob" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/cron-jobs/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "get": {
        "tags": ["Cron jobs"],
        "summary": "Get cron job by id",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmCronJob" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "tags": ["Cron jobs"],
        "summary": "Replace cron job by id",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrReplaceCronJobRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmCronJob" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "tags": ["Cron jobs"],
        "summary": "Partially update a cron job. When changing `agent_name`, the target must exist. Empty `updates` returns the current row.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatchCronJobRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmCronJob" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["Cron jobs"],
        "summary": "Delete cron job by id",
        "responses": {
          "204": { "description": "Deleted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/instances/register": {
      "post": {
        "tags": ["Instances"],
        "summary": "Register or upsert a Miroclaw runtime by instance_name",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterInstanceRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registered instance and revision_summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterInstanceResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/instances": {
      "get": {
        "tags": ["Instances"],
        "summary": "List runtime instances",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "description": "Filter by status (e.g. online)" }
          }
        ],
        "responses": {
          "200": {
            "description": "instances array",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instances": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/SwarmRuntimeInstance" }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/instances/{instance_name}": {
      "parameters": [
        {
          "name": "instance_name",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "tags": ["Instances"],
        "summary": "Get one runtime instance",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SwarmRuntimeInstance" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["Instances"],
        "summary": "Deregister runtime (offline + disable webhooks)",
        "responses": {
          "204": { "description": "Deregistered" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/instances/{instance_name}/heartbeat": {
      "parameters": [
        {
          "name": "instance_name",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "post": {
        "tags": ["Instances"],
        "summary": "Runtime heartbeat and health metadata",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HeartbeatRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "revision_summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revision_summary": { "$ref": "#/components/schemas/RevisionSummary" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/events/publish": {
      "post": {
        "tags": ["Events"],
        "summary": "Publish a change event (requires CLAWGOTCHA_INTERNAL_TOKEN)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PublishEventRequest" }
            }
          }
        },
        "responses": {
          "202": { "description": "Accepted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "description": "Forbidden" },
          "503": { "description": "Token not configured" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/events": {
      "get": {
        "tags": ["Events"],
        "summary": "Server-Sent Events stream of change notifications",
        "responses": {
          "200": {
            "description": "text/event-stream"
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Invalid id, body, or validation (e.g. `autonomy_level` not ReadOnly, Supervised, or Full; duplicate `name` where applicable; missing `agent` for `agent_name` on cron).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Not found (including GORM record not found)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalError": {
        "description": "Server error; detail is a generic message",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/PublicError" }
        }
      },
      "PublicError": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": { "type": "string", "description": "Machine-friendly code (e.g. `not_found`, `invalid id`, `internal_error`)" },
          "detail": { "type": "string", "description": "Optional human-readable detail" }
        }
      },
      "Healthz": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "ok" },
          "ts": { "type": "string", "format": "date-time", "description": "Server timestamp (RFC 3339 nanoseconds)" }
        }
      },
      "SwarmConfig": {
        "type": "object",
        "description": "GET/PUT responses use snake_case keys including default_provider, default_model, current_revision, and revision_summary.",
        "properties": {
          "default_provider": { "type": "string" },
          "default_model": { "type": "string" },
          "current_revision": { "type": "integer", "format": "int64" },
          "revision_summary": { "$ref": "#/components/schemas/RevisionSummary" }
        }
      },
      "PutConfigRequest": {
        "type": "object",
        "properties": {
          "default_provider": { "type": "string" },
          "default_model": { "type": "string" }
        }
      },
      "AutonomyLevel": {
        "type": "string",
        "description": "Required for create/put when supplied in handler validation",
        "enum": ["ReadOnly", "Supervised", "Full"]
      },
      "SwarmAgent": {
        "type": "object",
        "description": "PascalCase fields for most keys; `identity`, `soul`, `user_context` are the MiroClaw modular parts (also stored together in `SystemPrompt` with === markers). `modular_prompt` is true when `SystemPrompt` uses that format; when false, `user_context` mirrors the full legacy text for UIs that edit three parts.",
        "properties": {
          "ID": { "type": "string", "format": "uuid" },
          "Name": { "type": "string" },
          "SystemPrompt": { "type": "string" },
          "identity": { "type": "string" },
          "soul": { "type": "string" },
          "user_context": { "type": "string" },
          "modular_prompt": { "type": "boolean" },
          "Tools": { "type": "array", "items": { "type": "string" } },
          "Provider": { "type": "string" },
          "Model": { "type": "string" },
          "TimeoutSeconds": { "type": "integer" },
          "AutonomyLevel": { "$ref": "#/components/schemas/AutonomyLevel" },
          "current_revision": { "type": "integer", "format": "int64" },
          "last_changed_at": { "type": "string", "format": "date-time" },
          "deleted": { "type": "boolean" },
          "CreatedAt": { "type": "string", "format": "date-time" },
          "UpdatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateOrReplaceAgentRequest": {
        "type": "object",
        "description": "`name` and `autonomy_level` are required. If any of `identity`, `soul`, or `user_context` is present in the JSON, they are assembled (with MiroClaw === markers) into `system_prompt` in the DB; otherwise the flat `system_prompt` string is used.",
        "required": ["name", "autonomy_level"],
        "properties": {
          "name": { "type": "string" },
          "system_prompt": { "type": "string" },
          "identity": { "type": "string" },
          "soul": { "type": "string" },
          "user_context": { "type": "string" },
          "tools": { "type": "array", "items": { "type": "string" } },
          "provider": { "type": "string" },
          "model": { "type": "string" },
          "timeout_seconds": { "type": "integer" },
          "autonomy_level": { "$ref": "#/components/schemas/AutonomyLevel" }
        }
      },
      "PatchAgentRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "system_prompt": { "type": "string" },
          "identity": { "type": "string" },
          "soul": { "type": "string" },
          "user_context": { "type": "string" },
          "tools": { "type": "array", "items": { "type": "string" } },
          "clear_tools": { "type": "boolean", "description": "If true, forces `tools` to []." },
          "provider": { "type": "string" },
          "model": { "type": "string" },
          "timeout_seconds": { "type": "integer" },
          "autonomy_level": { "type": "string" }
        }
      },
      "AgentListResponse": {
        "type": "object",
        "required": ["agents", "revision_summary"],
        "properties": {
          "agents": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SwarmAgent" }
          },
          "revision_summary": { "$ref": "#/components/schemas/RevisionSummary" }
        }
      },
      "SwarmCronJob": {
        "type": "object",
        "description": "Response body uses Go default JSON (PascalCase) — `AgentName` not `agent_name`",
        "properties": {
          "ID": { "type": "string", "format": "uuid" },
          "Name": { "type": "string" },
          "AgentName": { "type": "string" },
          "Schedule": { "type": "string" },
          "TimeoutSeconds": { "type": "integer" },
          "Prompt": { "type": "string" },
          "Active": { "type": "boolean" },
          "CurrentRevision": { "type": "integer", "format": "int64" },
          "LastChangedAt": { "type": "string", "format": "date-time" },
          "CreatedAt": { "type": "string", "format": "date-time" },
          "UpdatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "RevisionSummary": {
        "type": "object",
        "properties": {
          "config_revision": { "type": "integer", "format": "int64" },
          "agents_max_revision": { "type": "integer", "format": "int64" },
          "cron_jobs_max_revision": { "type": "integer", "format": "int64" }
        }
      },
      "RegisterInstanceRequest": {
        "type": "object",
        "required": ["instance_name", "hostname", "version", "callback_url"],
        "properties": {
          "instance_name": { "type": "string" },
          "instance_type": { "type": "string", "default": "miroclaw" },
          "version": { "type": "string" },
          "hostname": { "type": "string" },
          "public_url": { "type": "string" },
          "callback_url": { "type": "string" },
          "capabilities": { "type": "array", "items": { "type": "string" } },
          "metadata": { "type": "object" },
          "started_at": { "type": "string", "format": "date-time" }
        }
      },
      "RegisterInstanceResponse": {
        "type": "object",
        "properties": {
          "instance": { "$ref": "#/components/schemas/SwarmRuntimeInstance" },
          "revision_summary": { "$ref": "#/components/schemas/RevisionSummary" }
        }
      },
      "SwarmRuntimeInstance": {
        "type": "object",
        "properties": {
          "ID": { "type": "string", "format": "uuid" },
          "InstanceName": { "type": "string" },
          "InstanceType": { "type": "string" },
          "Version": { "type": "string" },
          "Hostname": { "type": "string" },
          "PublicURL": { "type": "string" },
          "CallbackURL": { "type": "string" },
          "Capabilities": { "type": "array", "items": { "type": "string" } },
          "LastHeartbeatAt": { "type": "string", "format": "date-time" },
          "Status": { "type": "string" },
          "StartedAt": { "type": "string", "format": "date-time" },
          "Metadata": { "type": "object" },
          "CreatedAt": { "type": "string", "format": "date-time" },
          "UpdatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "HeartbeatRequest": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "metadata": { "type": "object" }
        }
      },
      "PublishEventRequest": {
        "type": "object",
        "required": ["event_type", "affected_entity_type"],
        "properties": {
          "event_type": { "type": "string" },
          "affected_entity_type": { "type": "string" },
          "affected_ids": { "type": "array", "items": { "type": "string" } },
          "new_revision": { "type": "integer", "format": "int64" }
        }
      },
      "CreateOrReplaceCronJobRequest": {
        "type": "object",
        "description": "Request body uses `json` struct tags: snake_case keys (`agent_name`, `timeout_seconds`, …). Responses use PascalCase (see `SwarmCronJob`).",
        "required": ["name", "agent_name"],
        "properties": {
          "name": { "type": "string" },
          "agent_name": { "type": "string" },
          "schedule": { "type": "string" },
          "timeout_seconds": { "type": "integer" },
          "prompt": { "type": "string" },
          "active": { "type": "boolean" }
        }
      },
      "PatchCronJobRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "agent_name": { "type": "string" },
          "schedule": { "type": "string" },
          "timeout_seconds": { "type": "integer" },
          "prompt": { "type": "string" },
          "active": { "type": "boolean" }
        }
      },
      "CronJobListResponse": {
        "type": "object",
        "required": ["cron_jobs", "revision_summary"],
        "properties": {
          "cron_jobs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SwarmCronJob" }
          },
          "revision_summary": { "$ref": "#/components/schemas/RevisionSummary" }
        }
      },
      "CronScheduleTimelineResponse": {
        "type": "object",
        "required": ["as_of", "horizon_ends", "anchored_by", "horizon_hours", "max_runs", "rows"],
        "properties": {
          "as_of": { "type": "string", "format": "date-time", "description": "Server time when the window starts" },
          "horizon_ends": { "type": "string", "format": "date-time", "description": "End of the projection window" },
          "anchored_by": { "type": "string", "example": "UpdatedAt" },
          "horizon_hours": { "type": "integer" },
          "max_runs": { "type": "integer" },
          "rows": { "type": "array", "items": { "$ref": "#/components/schemas/CronScheduleTimelineRow" } }
        }
      },
      "CronScheduleTimelineRow": {
        "type": "object",
        "description": "Mirrors a cron job; `ProjectedRuns` is empty if inactive, schedule is not standard cron, or no match in the window",
        "properties": {
          "ID": { "type": "string", "format": "uuid" },
          "Name": { "type": "string" },
          "AgentName": { "type": "string" },
          "Schedule": { "type": "string" },
          "Active": { "type": "boolean" },
          "UpdatedAt": { "type": "string", "format": "date-time" },
          "CreatedAt": { "type": "string", "format": "date-time" },
          "anchor_at": { "type": "string", "format": "date-time", "description": "Time used to anchor the first ticks (from `UpdatedAt` when set, else `CreatedAt`)" },
          "ScheduleParsed": { "type": "boolean" },
          "ParseError": { "type": "string" },
          "ProjectedRuns": { "type": "array", "items": { "type": "string", "format": "date-time" } }
        }
      }
    }
  }
}